HBASE-11437 Modify cell tag handling code to treat the length as unsigned. (Anoop)

+++ hbase-client/src/test/java/org/apache/hadoop/hbase/ipc/TestPayloadCarryingRpcController.java
@@ -171,7 +171,7 @@ public class TestPayloadCarryingRpcController {
               }
 
               @Override
-              public short getTagsLength() {
+              public int getTagsLength() {
                 // unused
                 return 0;
               }
